tests: add test for file and directory management
This test sets up multiple users and groups, multiple persistent storage
directories, and multiple files and directories under those storage
directories (some with overlapping destination paths), and then makes
assertions about the permissions on the source and destination paths.

NOTE that the permissions-related assertions are currently disabled.
This is an expedient to allow us to run the check before the intended
updates to the directory creation ordering and permissions-setting logic
are in place.