nixos: rework file and directory options
Adding the internal attributes:

    1. `prefix` - a path fragment prepended to `<file>.file` or
       `<directory>.dir` when expanding it relative to the persistent
       storage directory (exists to support prepending the implicit home
       directory in `users.<user>.files` and `users.<user>.directories`
       specifications),
    2. `relpath` - the file or directory path relative to the persistent
       storage path and `/`.
    3. `source` - the fully-qualified and normalized path rooted in the
       persistent storage path.
    4. `destination` - the fully-qualified and normalized path rooted in
        `/`.

This consolidates path-expansion logic into the file and directory
specifications themselves, and permits other parts of the Impermanence
module to refer to the expanded paths directly (most frequently,
`destination` and `source`).

Incidental change:  coerce strings to submodules in `files` and
`directories` definitions by defining the relevant options' types using
`lib.types.coercedTo` rather than by converting values with mkOption's
`apply` parameter.  Using type coercion rather than `apply` means that
the NixOS module system will automatically fill out the details for us
(e.g., applying option defaults for `persistentStoragePaths`, `prefix`,
`relpath`, and so on) so we do not have to explicitly construct a
complete submodule attrset ourselves.

BREAKING CHANGE: the type-coercion part of this changeset introduces a
liability to reduce the precedence of certain options associated with
persistent files and directories declared as strings rather than as
attrsets/submodules.

That is, files and directories declared like:

  {
    environment.persistence."/abc" = {
      files = [ "def" ];
      directories = [ "ghi" ];
    };
  }

Rather than:

  {
    environment.persistence."/abc" = {
      files = { file = "def"; <...>; };
      directories = { directory = "ghi"; <...>; };
    };
  }

In particular, the changes are liable to reduce the precedence of
options like `persistentStoragePath` and `parentDirectory` from the
default (the equivalent of the precedence associated with
`lib.mkOverride lib.modules.defaultPriority`) to the precedence
associated with `lib.mkOptionDefault`.