nixos: add `implicit` option on directories
This is a boolean field that indicates whether the directory was
explicitly defined in a `directories` list or whether it is a parent
directory of an explicitly-defined file.

If a directory is marked as implicit, then any of its associated
permissions settings are overridden by those associated with an
explicitly-defined directory with the same path.

This changeset additionally forbids inconsistent `parentDirectory`
permissions: Impermanence-managed persistent files that share a parent
directory must have identical `parentDirectory` mode, user, and group
settings.