nixos: no empty strings as file or directory names
BREAKING CHANGE: replace `lib.types.str` with `lib.types.nonEmptyStr` in
the type definitions for the `files.<name>.file` and
`directories.<name>.directory` options and their corresponding options
under `users.<user>`.

This prevents specifying the persistent storage directory itself (or
`<persistent-storage-directory>/<user-home>`) as a file or directory to
be linked or mounted into the installation target.